# analytics-foundation/analytics-foundation-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Analytics Foundation API
  version: 1.0.0
  description: >-
    Stable HTTP contract shared by deterministic application BFF paths and
    governed MCP tools. The API exposes data and workspace operations only; it
    does not own agent orchestration, prompts, human approval, or findings.
servers:
  - url: http://analytics-foundation-api:8200
paths:
  /health:
    get:
      operationId: health
      responses:
        '200':
          description: Process is alive.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HealthResponse'}
  /ready:
    get:
      operationId: ready
      responses:
        '200':
          description: Service is ready to accept requests.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HealthResponse'}
  /metadata:
    get:
      operationId: getDatasetMetadata
      responses:
        '200':
          description: Available Analytics Foundation datasets and tables.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DatasetMetadata'}
  /trends:
    get:
      operationId: getDisplayTrends
      parameters:
        - {$ref: '#/components/parameters/Days'}
        - {$ref: '#/components/parameters/StartDate'}
        - {$ref: '#/components/parameters/EndDate'}
        - {$ref: '#/components/parameters/LotIds'}
        - {$ref: '#/components/parameters/ProductIds'}
        - {$ref: '#/components/parameters/LayerIds'}
        - {$ref: '#/components/parameters/ExposureEquipmentIds'}
        - {$ref: '#/components/parameters/ChuckIds'}
      responses:
        '200':
          description: Trend series for deterministic application display.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TrendResponse'}
  /trends/query:
    post:
      operationId: queryTrends
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TrendQueryRequest'}
      responses:
        '200':
          description: Filtered trend series.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TrendResponse'}
  /trends/distribution:
    post:
      operationId: getTrendDistribution
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TrendQueryRequest'}
      responses:
        '200':
          description: Empirical KPI distribution statistics.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DistributionStats'}
  /workspaces:
    post:
      operationId: createWorkspace
      responses:
        '201':
          description: Workspace created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WorkspaceResponse'}
  /workspaces/{workspaceId}/filters:
    post:
      operationId: addWorkspaceFilters
      parameters:
        - {$ref: '#/components/parameters/WorkspaceId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WorkspaceFiltersRequest'}
      responses:
        '200':
          description: Filters associated with the workspace.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WorkspaceFiltersResponse'}
  /workspaces/{workspaceId}/connection-info:
    get:
      operationId: getWorkspaceConnectionInfo
      parameters:
        - {$ref: '#/components/parameters/WorkspaceId'}
      responses:
        '200':
          description: Opaque connection information for platform-side consumers.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WorkspaceConnectionInfo'}
  /workspaces/{workspaceId}/registrations:
    post:
      operationId: registerDataset
      parameters:
        - {$ref: '#/components/parameters/WorkspaceId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegistrationRequest'}
      responses:
        '202':
          description: Registration accepted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RegistrationStatus'}
  /workspaces/{workspaceId}/registrations/{registrationId}:
    get:
      operationId: getRegistrationStatus
      parameters:
        - {$ref: '#/components/parameters/WorkspaceId'}
        - name: registrationId
          in: path
          required: true
          schema: {type: string, minLength: 1}
      responses:
        '200':
          description: Current registration status.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RegistrationStatus'}
  /wafers/query:
    post:
      operationId: queryWafers
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WaferQueryRequest'}
      responses:
        '200':
          description: Wafer query result.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WaferQueryResponse'}
  /tdbb/runs:
    post:
      operationId: runTdbb
      description: >-
        Request TDBB processing (default 10par model, AVG and W2W context
        levels) for all lots matching the filters before and after
        change_date. The mock resolves the request to completed TDBB runs that
        already exist (one run per lot) and returns their IDs with budgets and
        wafer/field maps per period.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TdbbRunRequest'}
      responses:
        '201':
          description: Completed TDBB runs with budgets and wafer/field maps per period.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TdbbRunResult'}
        '422':
          description: Invalid run request, e.g. change_date outside the window.
  /tdbb/runs/{runId}:
    get:
      operationId: getTdbbRun
      parameters:
        - {$ref: '#/components/parameters/RunId'}
      responses:
        '200':
          description: TDBB run metadata.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TdbbRunInfo'}
        '404':
          description: Unknown run.
  /tdbb/runs/{runId}/data:
    get:
      operationId: getTdbbData
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - name: tables
          in: query
          style: form
          explode: true
          description: Subset of TDBB tables; all tables when omitted.
          schema:
            type: array
            items: {type: string, enum: [ce_wafer, ce_field, nce_wafer, nce_field]}
      responses:
        '200':
          description: TDBB output rows of one run, keyed by table.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TdbbRunData'}
        '404':
          description: Unknown run.
        '422':
          description: Unknown table.
components:
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema: {type: string, minLength: 1}
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: {type: string, minLength: 1}
    Days:
      name: days
      in: query
      schema: {type: integer, minimum: 1, maximum: 90}
    StartDate:
      name: start_date
      in: query
      schema: {type: string, format: date}
    EndDate:
      name: end_date
      in: query
      schema: {type: string, format: date}
    LotIds:
      name: lot_ids
      in: query
      style: form
      explode: true
      schema: {type: array, items: {type: string}}
    ProductIds:
      name: product_ids
      in: query
      style: form
      explode: true
      schema: {type: array, items: {type: string}}
    LayerIds:
      name: layer_ids
      in: query
      style: form
      explode: true
      schema: {type: array, items: {type: string}}
    ExposureEquipmentIds:
      name: exposure_equipment_ids
      in: query
      style: form
      explode: true
      schema: {type: array, items: {type: string}}
    ChuckIds:
      name: chuck_ids
      in: query
      style: form
      explode: true
      schema: {type: array, items: {type: string}}
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [ok]}
      additionalProperties: false
    DatasetMetadata:
      type: object
      required: [trend_table, wafer_table]
      properties:
        trend_table: {type: string, minLength: 1}
        wafer_table: {type: string, minLength: 1}
      additionalProperties: true
    TrendQueryRequest:
      type: object
      properties:
        days: {type: integer, minimum: 1, maximum: 90, nullable: true}
        start_date: {type: string, format: date, nullable: true}
        end_date: {type: string, format: date, nullable: true}
        lot_ids: {type: array, items: {type: string}, default: []}
        product_ids: {type: array, items: {type: string}, default: []}
        layer_ids: {type: array, items: {type: string}, default: []}
        exposure_equipment_ids: {type: array, items: {type: string}, default: []}
        chuck_ids: {type: array, items: {type: string}, default: []}
      additionalProperties: false
    TrendResponse:
      type: object
      required: [series]
      properties:
        series:
          type: array
          items: {$ref: '#/components/schemas/TrendSeries'}
      additionalProperties: false
    TrendSeries:
      type: object
      required: [machine, product, points]
      properties:
        machine: {type: string}
        product: {type: string}
        lot_id: {type: string, nullable: true}
        layer_id: {type: string, nullable: true}
        exposure_equipment_id: {type: string, nullable: true}
        points:
          type: array
          items: {$ref: '#/components/schemas/TrendPoint'}
      additionalProperties: false
    TrendPoint:
      type: object
      required: [date, kpi_value]
      properties:
        date: {type: string, format: date-time}
        kpi_value: {type: number, description: Overlay X KPI (|mean|+3 sigma) in nm.}
        kpi_value_y: {type: number, nullable: true, description: Overlay Y KPI (|mean|+3 sigma) in nm.}
        lot_id: {type: string, nullable: true}
        wafer_id: {type: string, nullable: true}
        chuck_id: {type: string, nullable: true}
      additionalProperties: false
    DistributionStats:
      type: object
      required: [sample_count]
      properties:
        sample_count: {type: integer, minimum: 0}
        p95: {type: number, nullable: true}
        p99: {type: number, nullable: true}
        mean: {type: number, nullable: true}
        stdev: {type: number, nullable: true}
        bell_curve_range:
          type: object
          nullable: true
          properties:
            lower: {type: number}
            upper: {type: number}
          required: [lower, upper]
          additionalProperties: false
      additionalProperties: false
    WorkspaceResponse:
      type: object
      required: [workspace_id]
      properties:
        workspace_id: {type: string, minLength: 1}
      additionalProperties: false
    WorkspaceFiltersRequest:
      type: object
      required: [filters]
      properties:
        filters: {type: object, additionalProperties: true}
      additionalProperties: false
    WorkspaceFiltersResponse:
      type: object
      required: [workspace_id, filters]
      properties:
        workspace_id: {type: string}
        filters: {type: object, additionalProperties: true}
      additionalProperties: false
    WorkspaceConnectionInfo:
      type: object
      required: [workspace_id, values]
      properties:
        workspace_id: {type: string}
        values: {type: object, additionalProperties: true}
      additionalProperties: false
    RegistrationRequest:
      type: object
      required: [dataset, table]
      properties:
        dataset: {type: string, minLength: 1}
        table: {type: string, minLength: 1}
      additionalProperties: false
    RegistrationStatus:
      type: object
      required: [registration_id, workspace_id, status, progress_pct, table]
      properties:
        registration_id: {type: string, minLength: 1}
        workspace_id: {type: string, minLength: 1}
        status: {type: string, enum: [PENDING, RUNNING, READY, FAILED]}
        progress_pct: {type: integer, minimum: 0, maximum: 100}
        table: {type: string, minLength: 1}
        error: {type: string, nullable: true}
      additionalProperties: false
    WaferQueryRequest:
      type: object
      required: [workspace_id, table]
      properties:
        workspace_id: {type: string, minLength: 1}
        table: {type: string, minLength: 1}
        filters: {type: object, additionalProperties: true, default: {}}
      additionalProperties: false
    WaferQueryResponse:
      type: object
      required: [workspace_id, table, rows, anomalous_wafers]
      properties:
        workspace_id: {type: string}
        table: {type: string}
        rows:
          type: array
          items: {type: object, additionalProperties: true}
        anomalous_wafers:
          type: array
          items: {type: string}
      additionalProperties: false
    TdbbRunRequest:
      type: object
      required: [start_date, end_date, change_date]
      properties:
        start_date: {type: string, format: date}
        end_date: {type: string, format: date}
        change_date: {type: string, format: date, description: First day of the after period.}
        lot_ids: {type: array, items: {type: string}, default: []}
        product_ids: {type: array, items: {type: string}, default: []}
        layer_ids: {type: array, items: {type: string}, default: []}
        exposure_equipment_ids: {type: array, items: {type: string}, default: []}
        chuck_ids: {type: array, items: {type: string}, default: []}
        model_step: {type: string, enum: [10par], default: 10par}
        context_levels:
          type: array
          items: {type: string, enum: [AVG, W2W]}
          default: [AVG, W2W]
      additionalProperties: false
    TdbbRunResult:
      type: object
      required: [status, change_date, settings, periods, maps]
      properties:
        status: {type: string, enum: [COMPLETED]}
        change_date: {type: string, format: date}
        settings:
          type: object
          required: [model_step, context_levels, budgets]
          properties:
            model_step: {type: string}
            context_levels: {type: array, items: {type: string}}
            budgets: {type: array, items: {type: string}}
          additionalProperties: false
        periods:
          type: array
          items: {$ref: '#/components/schemas/TdbbPeriod'}
        maps:
          type: array
          items: {$ref: '#/components/schemas/TdbbMap'}
      additionalProperties: false
    TdbbPeriod:
      type: object
      required: [period, start_date, end_date, run_ids, lot_count, wafer_count, budgets]
      properties:
        period: {type: string, enum: [before, after]}
        start_date: {type: string, format: date}
        end_date: {type: string, format: date}
        run_ids:
          type: array
          description: Completed TDBB runs (one per lot) used for this period.
          items: {type: string}
        lot_count: {type: integer, minimum: 0}
        wafer_count: {type: integer, minimum: 0}
        budgets:
          type: array
          items:
            type: object
            required: [budget, label, metric, metric_label, context, context_label]
            properties:
              budget: {type: string, description: "<metric>.<context>, e.g. nce_wafer.average"}
              label: {type: string}
              metric: {type: string, enum: [nce_wafer, nce_field, ce_wafer, ce_field, ce_translation]}
              metric_label: {type: string}
              context: {type: string, enum: [average, chuck_to_chuck, lot_to_lot, wafer_to_wafer]}
              context_label: {type: string}
              x_m3s: {type: number, nullable: true, description: nm}
              y_m3s: {type: number, nullable: true, description: nm}
            additionalProperties: false
      additionalProperties: false
    TdbbMap:
      type: object
      required: [budget, period, level, points]
      properties:
        budget: {type: string}
        period: {type: string, enum: [before, after]}
        level: {type: string, enum: [wafer, field], description: Averaged per field center or per intrafield position.}
        points:
          type: array
          items:
            type: object
            required: [x, y, dx, dy]
            properties:
              x: {type: number, description: mm}
              y: {type: number, description: mm}
              dx: {type: number, description: Mean X in nm}
              dy: {type: number, description: Mean Y in nm}
              m3s_x: {type: number, nullable: true}
              m3s_y: {type: number, nullable: true}
            additionalProperties: false
      additionalProperties: false
    TdbbRunInfo:
      type: object
      required: [run_id, status, model_step, context_levels, product_id, layer_id, exposure_equipment_id, lot_id, lot_start, wafers]
      properties:
        run_id: {type: string, minLength: 1}
        status: {type: string, enum: [COMPLETED]}
        model_step: {type: string}
        context_levels: {type: array, items: {type: string}}
        product_id: {type: string}
        layer_id: {type: string}
        exposure_equipment_id: {type: string}
        lot_id: {type: string}
        lot_start: {type: string, format: date-time}
        wafers:
          type: array
          items:
            type: object
            required: [wafer_id]
            properties:
              wafer_id: {type: string}
              chuck_id: {type: string, nullable: true}
            additionalProperties: false
      additionalProperties: false
    TdbbRunData:
      type: object
      required: [run_id, tables]
      properties:
        run_id: {type: string, minLength: 1}
        tables:
          type: object
          description: >-
            Rows per TDBB table (ce_wafer, ce_field, nce_wafer, nce_field) in the
            TDBB output schema; context_level AVG per lot, W2W per wafer.
          additionalProperties:
            type: array
            items: {type: object, additionalProperties: true}
      additionalProperties: false